Restructure the files so that jit-preloader takes Rails 6 as the current release
