+++ lib/jit_preloader/active_record/associations/preloader/ar5_association.rb
@@ -5,31 +5,33 @@ module JitPreloader
     # below. Our changes here are that we remove records that are already
     # part of the target, then attach all of the records to a new jit preloader.
     #
-    # def run
-    #   if !preload_scope || preload_scope.empty_scope?
-    #     owners.each do |owner|
-    #       associate_records_to_owner(owner, records_by_owner[owner] || [])
-    #     end
-    #   else
-    #     # Custom preload scope is used and
-    #     # the association can not be marked as loaded
-    #     # Loading into a Hash instead
-    #     records_by_owner
+    # def run(preloader)
+    #   records = load_records do |record|
+    #     owner = owners_by_key[convert_key(record[association_key_name])]
+    #     association = owner.association(reflection.name)
+    #     association.set_inverse_instance(record)
     #   end
-    #   self
+
+    #   owners.each do |owner|
+    #     associate_records_to_owner(owner, records[convert_key(owner[owner_key_name])] || [])
+    #   end
     # end
-    def run
+
+    def run(preloader)
       all_records = []
+      records = load_records do |record|
+        owner = owners_by_key[convert_key(record[association_key_name])]
+        association = owner.association(reflection.name)
+        association.set_inverse_instance(record)
+      end
 
       owners.each do |owner|
-        owned_records = records_by_owner[owner] || []
+        owned_records = records[convert_key(owner[owner_key_name])] || []
         all_records.concat(Array(owned_records)) if owner.jit_preloader || JitPreloader.globally_enabled?
         associate_records_to_owner(owner, owned_records)
       end
 
       JitPreloader::Preloader.attach(all_records) if all_records.any?
-
-      self
     end
 
     # Original method:
